Refactored utility functions common to vnode and vserver
